viz/requirements.txt
annotated-types==0.6.0
ansi2html==1.9.1
anyio==3.7.1
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
exceptiongroup==1.2.0
fastapi==0.105.0
Flask==3.0.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
more-itertools==4.3.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
plotly==5.18.0
pluggy==0.8.0
py==1.7.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==2.3.0
pytest==4.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
six==1.11.0
sniffio==1.3.0
starlette==0.27.0
svgwrite==1.2.1
tabulate==0.9.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
visdcc==0.0.50
Werkzeug==3.0.1
zipp==3.17.0
antlr4-python3-runtime==4.11.1
PyYAML